+++ src/app/services/alerts/alerts.service.ts
@@ -4,6 +4,7 @@ import { firstValueFrom } from 'rxjs';
 
 import { AppModalsService } from '../modals/app-modals.service';
 import { AppErrorAlertData } from './alerts.model';
+import { AppModalConfig } from '../modals/app-modals.model';
 
 @Injectable({
   providedIn: 'root'
@@ -15,10 +16,10 @@ export class AlertsService {
     private modalsService: AppModalsService,
   ) {}
 
-  async showErrorAlert(errorText: string = this.defaultError): Promise<void> {
+  async showErrorAlert(errorText: string = this.defaultError, config: AppModalConfig = {}): Promise<void> {
     const { ErrorAlertComponent } = await import('../../modals/error-alert/error-alert.component');
     await firstValueFrom(
-      this.modalsService.openModal<unknown, AppErrorAlertData>(ErrorAlertComponent, { errorText }, {})
+      this.modalsService.openModal<unknown, AppErrorAlertData>(ErrorAlertComponent, { errorText }, { ...config })
     );
   }
 }